emsmdb: resolve cov-scan warning about improper printf
CID 1592968: (#3 of 3): Invalid type in argument to printf format specifier (PRINTF_ARGS)
invalid_type: Argument result to format specifier %x was expected to
have type `unsigned int` but has type `ec_error_t`.
CID 1592969: (#3 of 3): Printf arg type mismatch (PW.PRINTF_ARG_MISMATCH)
1. printf_arg_mismatch: argument is incompatible with corresponding
format string conversion (expected type "unsigned int" but argument has
type "ec_error_t")

Harmless in practice because both are size 4 on contemporary platforms.

Fixes: gromox-2.2-6-gc996f0a08

+++ exch/emsmdb/rop_processor.cpp
@@ -669,7 +669,8 @@ ec_error_t rop_processor_proc(uint32_t flags, const uint8_t *pin,
 			         &tmp_cb, &rop_buff, false, response_list);
 			if (g_rop_debug >= 2 || (g_rop_debug >= 1 && result != ecSuccess))
 				mlog(LV_DEBUG, "rop_proc_ex+chain() %luµs EC=%xh",
-					msec(tp_now() - rlog_start), result);
+					msec(tp_now() - rlog_start),
+					static_cast<unsigned int>(result));
 			if (result != ecSuccess)
 				break;
 			if (response_list.empty())
@@ -695,7 +696,8 @@ ec_error_t rop_processor_proc(uint32_t flags, const uint8_t *pin,
 			         &tmp_cb, &rop_buff, false, response_list);
 			if (g_rop_debug >= 2 || (g_rop_debug >= 1 && result != ecSuccess))
 				mlog(LV_DEBUG, "rop_proc_ex+chain() %luµs EC=%xh",
-					msec(tp_now() - rlog_start), result);
+					msec(tp_now() - rlog_start),
+					static_cast<unsigned int>(result));
 			if (result != ecSuccess)
 				break;
 			if (response_list.empty())
@@ -722,7 +724,8 @@ ec_error_t rop_processor_proc(uint32_t flags, const uint8_t *pin,
 			         &tmp_cb, &rop_buff, false, response_list);
 			if (g_rop_debug >= 2 || (g_rop_debug >= 1 && result != ecSuccess))
 				mlog(LV_DEBUG, "rop_proc_ex+chain() %luµs EC=%xh",
-					msec(tp_now() - rlog_start), result);
+					msec(tp_now() - rlog_start),
+					static_cast<unsigned int>(result));
 			if (result != ecSuccess)
 				break;
 			if (response_list.empty())
